these need to be doubles so they don't overflow
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@27258 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/proto/e_phys/src/lib/E_Phys.h
@@ -169,10 +169,10 @@ struct _E_Phys_World
 
   Ecore_Timer *timer;
 
-  float time;
+  double time;
 
-  float dt;
-  float elapsed;
+  double dt;
+  double elapsed;
   int constraint_iter;
   float friction;
