+++ e17/proto/e_phys/src/lib/e_phys_world.c
@@ -45,13 +45,9 @@ static int
 e_phys_world_timer(void *data)
 {
   E_Phys_World *world;
-  float now;
+  double now;
   int i;
 
-  // XXX right now, each timeslice is world->dt apart. it would probably
-  // be better to get the current clock time and figure out how far (real time)
-  // actually elapsed between slices. (then if a slice takes longer than dt to
-  // calc, the frame rate will drop, the physical motion will remain 'at speed'
   world = data;
   now = ecore_time_get();
   world->elapsed = now - world->time;
